refactor(storage): cleanup stale buckets code moves
Move the functions to remove stale buckets to a common library so I can
use it to remove the stale buckets created by any integration tests and
benchmarks.
