+++ google/cloud/storage/testing/remove_stale_buckets_test.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "google/cloud/storage/examples/storage_examples_common.h"
+#include "google/cloud/storage/testing/remove_stale_buckets.h"
 #include "google/cloud/storage/testing/mock_client.h"
 #include "google/cloud/internal/format_time_point.h"
 #include "google/cloud/testing_util/status_matchers.h"
@@ -20,7 +20,7 @@
 namespace google {
 namespace cloud {
 namespace storage {
-namespace examples {
+namespace testing {
 namespace {
 
 using ::google::cloud::testing_util::StatusIs;
@@ -115,7 +115,7 @@ TEST(CleanupStaleBucketsTest, RemoveStaleBuckets) {
 }
 
 }  // namespace
-}  // namespace examples
+}  // namespace testing
 }  // namespace storage
 }  // namespace cloud
 }  // namespace google
